Znaczące zmiany

+++ SmartHouseAppServer/SmartHouseApp.Share/ViewModel/StaticRouterDataViewModel.cs
@@ -13,8 +13,8 @@ namespace SmartHouseApp.Share.ViewModel
         public double TrasmitterPower { get; set; }
         public double AntennaGain { get; set; }
         public double FadeMargin { get; set; }
-        public decimal LocationX { get; set; }
-        public decimal LocationY { get; set; }
-        public decimal LocationZ { get; set; }
+        public double LocationX { get; set; }
+        public double LocationY { get; set; }
+        public double LocationZ { get; set; }
     }
 }
